Consolidating ruleset files. Cleaning up warnings
