+++ test/WebJobs.Script.Tests.Shared/WebJobs.Script.Tests.Shared.projitems
@@ -19,7 +19,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)TestHostBuilderExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestHostExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestNameResolver.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)TestScriptJobHostEnvironment.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestTraits.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestHandler.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestHelpers.cs" />
