--- test/test.ruleset
@@ -1,45 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<RuleSet Name="WebJobs SDK Test" Description="This rule set contains the subset of rules that applies to tests. " ToolsVersion="12.0">
-  <IncludeAll Action="Error" />
-  <Rules AnalyzerId="Microsoft.Analyzers.ManagedCodeAnalysis" RuleNamespace="Microsoft.Rules.Managed">
-    <Rule Id="CA1001" Action="None" />
-    <Rule Id="CA1002" Action="None" />
-    <Rule Id="CA1004" Action="None" />
-    <Rule Id="CA1005" Action="None" />
-    <Rule Id="CA1006" Action="None" />
-    <Rule Id="CA1011" Action="None" />
-    <Rule Id="CA1020" Action="None" />
-    <Rule Id="CA1024" Action="None" />
-    <Rule Id="CA1026" Action="None" />
-    <Rule Id="CA1031" Action="None" />
-    <Rule Id="CA1033" Action="None" />
-    <Rule Id="CA1034" Action="None" />
-    <Rule Id="CA1040" Action="None" />
-    <Rule Id="CA1062" Action="None" />
-    <Rule Id="CA1065" Action="None" />
-    <Rule Id="CA1303" Action="None" />
-    <Rule Id="CA1305" Action="None" />
-    <Rule Id="CA1307" Action="None" />
-    <Rule Id="CA1506" Action="None" />
-    <Rule Id="CA1707" Action="None" />
-    <Rule Id="CA1720" Action="None" />
-    <Rule Id="CA1726" Action="None" />
-    <Rule Id="CA1806" Action="None" />
-    <Rule Id="CA1811" Action="None" />
-    <Rule Id="CA1813" Action="None" />
-    <Rule Id="CA1819" Action="None" />
-    <Rule Id="CA1822" Action="None" />
-    <Rule Id="CA2000" Action="None" />
-    <Rule Id="CA2201" Action="None" />
-    <Rule Id="CA2210" Action="None" />
-    <Rule Id="CA2234" Action="None" />
-    <Rule Id="CA2237" Action="None" />
-    <Rule Id="CA903" Action="None" />
-    <Rule Id="CA904" Action="None" />
-    <Rule Id="CA905" Action="None" />
-    <Rule Id="CA908" Action="None" />
-    <Rule Id="CA2100" Action="None" />
-    <Rule Id="CA1054" Action="None" />
-    <Rule Id="CS0618" Action="None" /> <!-- Ignore Obsolete attributes for 2.0 release -->
-  </Rules>
-</RuleSet>
